chore: 프로젝트 패키지 재구조화
