+++ src/main/java/com/hanghae/bulletbox/daily/dto/RequestLoadFavoriteDto.java
@@ -1,4 +1,4 @@
-package com.hanghae.bulletbox.todo.dto;
+package com.hanghae.bulletbox.daily.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
