+++ src/main/java/com/hanghae/bulletbox/daily/dto/RequestUpdateTodoDto.java
@@ -1,4 +1,6 @@
-package com.hanghae.bulletbox.todo.dto;
+package com.hanghae.bulletbox.daily.dto;
+
+import com.hanghae.bulletbox.todo.dto.TodoMemoDto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 
@@ -10,7 +12,11 @@ import java.util.List;
 
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class RequestCreateTodoDto {
+@Schema(description = "할 일 수정 요청 Dto")
+public class RequestUpdateTodoDto {
+
+    @Schema(description = "할 일 ID", type = "Long", example = "1")
+    private Long todoId;
 
     @Schema(description = "할 일 내용", type = "String", example = "할 일 내용")
     private String todoContent;
@@ -38,4 +44,5 @@ public class RequestCreateTodoDto {
 
     @Schema(description = "할 일의 메모 변수", type = "List")
     private List<TodoMemoDto> memos;
+
 }
