+++ src/main/java/com/hanghae/bulletbox/daily/dto/ResponseLoadFavoriteDto.java
@@ -1,6 +1,8 @@
-package com.hanghae.bulletbox.todo.dto;
+package com.hanghae.bulletbox.daily.dto;
 
 import com.hanghae.bulletbox.todo.TodoBullet;
+import com.hanghae.bulletbox.todo.dto.TodoDto;
+import com.hanghae.bulletbox.todo.dto.TodoMemoDto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
