+++ src/main/java/com/hanghae/bulletbox/daily/dto/ResponseShowTodoCreatePageDto.java
@@ -1,4 +1,4 @@
-package com.hanghae.bulletbox.todo.dto;
+package com.hanghae.bulletbox.daily.dto;
 
 import com.hanghae.bulletbox.category.dto.CategoryDto;
