+++ src/main/java/com/hanghae/bulletbox/main/dto/ResponseShowDailyByCategoryDto.java
@@ -1,4 +1,6 @@
-package com.hanghae.bulletbox.todo.dto;
+package com.hanghae.bulletbox.main.dto;
+
+import com.hanghae.bulletbox.daily.dto.DailyDto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
