+++ src/main/java/com/hanghae/bulletbox/main/dto/ResponseShowDailyDto.java
@@ -1,8 +1,9 @@
-package com.hanghae.bulletbox.todo.dto;
+package com.hanghae.bulletbox.main.dto;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 
 import com.hanghae.bulletbox.category.dto.CategoryDto;
+import com.hanghae.bulletbox.daily.dto.DailyDto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
