+++ src/main/java/com/hanghae/bulletbox/main/dto/ResponseShowMainPageDto.java
@@ -1,6 +1,7 @@
-package com.hanghae.bulletbox.todo.dto;
+package com.hanghae.bulletbox.main.dto;
 
 import com.hanghae.bulletbox.category.dto.CategoryDto;
+import com.hanghae.bulletbox.daily.dto.DailyDto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
